requirements.txt
asn1crypto==0.24.0
boto3==1.9.156
botocore==1.12.156
certifi==2019.3.9
cffi==1.12.3
chardet==3.0.4
cryptography==2.6.1
docutils==0.14
entrypoints==0.3
future==0.17.1
idna==2.8
jmespath==0.9.4
keyring==19.0.2
knockknock==0.1.3
numpy==1.16.3
Pillow==6.0.0
pycparser==2.19
python-dateutil==2.8.0
python-dotenv==0.10.2
python-telegram-bot==11.1.0
pytorch-pretrained-bert==0.6.2
regex==2019.5.25
requests==2.22.0
s3transfer==0.2.0
six==1.12.0
torch==1.1.0
torchvision==0.3.0
tqdm==4.32.1
urllib3==1.25.3
yagmail==0.11.220
